+++ src/refactor_nrepl/ns/libspec_allowlist.clj
@@ -6,20 +6,33 @@
    (clojure.lang IFn)
    (java.util.regex Pattern)))
 
-(defn- libspec-allowlist* []
-  (let [kondo-file (io/file ".clj-kondo" "config.edn")
-        exclude (when (.exists kondo-file)
-                  (try
-                    (-> kondo-file slurp read-string :linters :unused-namespace :exclude)
-                    (catch Exception e
-                      (when (System/getenv "CI")
-                        (throw e)))))]
-    (->> exclude
-         (mapv (fn [entry]
-                 (if (symbol? entry)
-                   (str "^" (Pattern/quote (str entry)) "$")
-                   entry)))
-         (into (:libspec-whitelist config/*config*)))))
+(defn maybe-unwrap-quote [obj]
+  (if (and (seq? obj) (= 'quote (first obj)))
+    (second obj)
+    obj))
+
+(defn- kondo-excludes [{namespace-name :ns
+                        ns-meta        :meta}]
+  (let [linter-path [:linters :unused-namespace :exclude]
+        local-config (-> ns-meta :clj-kondo/config maybe-unwrap-quote)
+        kondo-file (io/file ".clj-kondo" "config.edn")
+        config (when (.exists kondo-file)
+                 (try
+                   (-> kondo-file slurp read-string)
+                   (catch Exception e
+                     (when (System/getenv "CI")
+                       (throw e)))))]
+    (reduce into [(get-in config linter-path)
+                  (get-in config (into [:config-in-ns namespace-name] linter-path))
+                  (get-in local-config linter-path)])))
+
+(defn- libspec-allowlist* [current-ns]
+  (->> (kondo-excludes current-ns)
+       (mapv (fn [entry]
+               (if (symbol? entry)
+                 (str "^" (Pattern/quote (str entry)) "$")
+                 entry)))
+       (into (:libspec-whitelist config/*config*))))
 
 (def ^:private ^:dynamic ^IFn *libspec-allowlist* nil)
 
@@ -32,9 +45,9 @@
   with clj-kondo's `:unused-namespace` config.
 
   Uses a memoized version if available."
-  []
-  (or (some-> *libspec-allowlist* .invoke)
-      (libspec-allowlist*)))
+  [current-ns]
+  (or (some-> *libspec-allowlist* (.invoke current-ns))
+      (libspec-allowlist* current-ns)))
 
 (defmacro with-memoized-libspec-allowlist
   "Memoizes the libspec-allowlist internals while `body` is executing.