+++ src/refactor_nrepl/ns/ns_parser.clj
@@ -112,12 +112,16 @@
          (parse-clj-or-cljs-ns path :cljs)))
 
 (defn parse-ns [path-or-file]
-  (assoc
-   (if (core/cljc-file? (io/file path-or-file))
-     (parse-cljc-ns path-or-file)
-     (parse-clj-or-cljs-ns path-or-file))
-   :ns (second (core/read-ns-form-with-meta path-or-file))
-   :source-dialect (core/file->dialect path-or-file)))
+  (let [[_ namespace-name :as ns-form] (core/read-ns-form-with-meta path-or-file)
+        select-metadata (juxt :top-level-meta :attr-map)
+        metadata (-> ns-form meta select-metadata)]
+    (assoc
+     (if (core/cljc-file? (io/file path-or-file))
+       (parse-cljc-ns path-or-file)
+       (parse-clj-or-cljs-ns path-or-file))
+     :ns namespace-name
+     :meta (apply merge metadata)
+     :source-dialect (core/file->dialect path-or-file))))
 
 (def ^:dynamic *read-ns-form-with-meta* core/read-ns-form-with-meta)